+++ src/App.tsx
@@ -1,324 +1,37 @@
-import React, { useState, useCallback, useMemo, useEffect } from "react";
-import { generateItems, renderLog } from "./utils";
-import { ThemeProvider, useTheme } from "./contexts/ThemeContext";
-import { AuthProvider, useAuth } from "./contexts/AuthContext";
+import React, { useState } from "react";
 import {
-  NotificationProvider,
-  useNotification
-} from "./contexts/notificationContext";
-
-// 타입 정의
-interface Item {
-  id: number;
-  name: string;
-  category: string;
-  price: number;
-}
-
-// 별도의 컴포넌트로 notification 로직 분리
-const AuthNotification: React.FC<{ action: string }> = React.memo(
-  ({ action }) => {
-    const { addNotification } = useNotification();
-
-    useEffect(() => {
-      if (action === "login") {
-        addNotification("성공적으로 로그인되었습니다", "success");
-      } else if (action === "logout") {
-        addNotification("로그아웃되었습니다", "info");
-      }
-    }, [action, addNotification]);
-
-    return null;
-  }
-);
-
-// Header 컴포넌트
-// NOTE: 알림 추가 및 닫기 시 header도 같이 리렌더링 되어 React.memo를 사용하여 최적화 -> 그래도 안됨.
-export const Header: React.FC = React.memo(() => {
-  renderLog("Header rendered");
-  const { theme, toggleTheme } = useTheme();
-  const { user, login, logout } = useAuth();
-  const [authAction, setAuthAction] = useState<string>(""); // 추가
-
-  const handleLogin = useCallback(() => {
-    // 실제 애플리케이션에서는 사용자 입력을 받아야 합니다.
-    login("user@example.com", "password");
-    setAuthAction("login");
-  }, [login]);
-
-  const handleLogout = useCallback(() => {
-    logout();
-    setAuthAction("logout");
-  }, [logout]);
-
-  return (
-    <>
-      <header className="bg-gray-800 text-white p-4">
-        <div className="container mx-auto flex justify-between items-center">
-          <h1 className="text-2xl font-bold">샘플 애플리케이션</h1>
-          <div className="flex items-center">
-            <button
-              onClick={toggleTheme}
-              className="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded mr-2"
-            >
-              {theme === "light" ? "다크 모드" : "라이트 모드"}
-            </button>
-            {user ? (
-              <div className="flex items-center">
-                <span className="mr-2">{user.name}님 환영합니다!</span>
-                <button
-                  onClick={handleLogout}
-                  className="bg-red-500 hover:bg-red-700 text-white font-bold py-2 px-4 rounded"
-                >
-                  로그아웃
-                </button>
-              </div>
-            ) : (
-              <button
-                onClick={handleLogin}
-                className="bg-green-500 hover:bg-green-700 text-white font-bold py-2 px-4 rounded"
-              >
-                로그인
-              </button>
-            )}
-          </div>
-        </div>
-      </header>
-      <AuthNotification action={authAction} />
-    </>
-  );
-});
-
-// ItemList 컴포넌트
-export const ItemList: React.FC<{
-  items: Item[];
-  onAddItemsClick: () => void;
-}> = React.memo(
-  ({ items, onAddItemsClick }) => {
-    renderLog("ItemList rendered");
-    const [filter, setFilter] = useState("");
-    const { theme } = useTheme();
-
-    const filteredItems = useMemo(
-      () =>
-        items.filter(
-          (item) =>
-            item.name.toLowerCase().includes(filter.toLowerCase()) ||
-            item.category.toLowerCase().includes(filter.toLowerCase())
-        ),
-      [items, filter]
-    );
-
-    const { totalPrice, averagePrice } = useMemo(() => {
-      const total = filteredItems.reduce((sum, item) => sum + item.price, 0);
-      return {
-        totalPrice: total,
-        averagePrice: Math.round(total / filteredItems.length) || 0
-      };
-    }, [filteredItems]);
-
-    return (
-      <div className="mt-8">
-        <div className="flex justify-between items-center mb-4">
-          <h2 className="text-2xl font-bold">상품 목록</h2>
-          <div>
-            <button
-              type="button"
-              className="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded text-xs"
-              onClick={onAddItemsClick}
-            >
-              대량추가
-            </button>
-          </div>
-        </div>
-        <input
-          type="text"
-          placeholder="상품 검색..."
-          value={filter}
-          onChange={(e) => setFilter(e.target.value)}
-          className="w-full p-2 mb-4 border border-gray-300 rounded text-black"
-        />
-        <ul className="mb-4 mx-4 flex gap-3 text-sm justify-end">
-          <li>검색결과: {filteredItems.length.toLocaleString()}개</li>
-          <li>전체가격: {totalPrice.toLocaleString()}원</li>
-          <li>평균가격: {averagePrice.toLocaleString()}원</li>
-        </ul>
-        <ul className="space-y-2">
-          {filteredItems.map((item, index) => (
-            <li
-              key={index}
-              className={`p-2 rounded shadow ${theme === "light" ? "bg-white text-black" : "bg-gray-700 text-white"}`}
-            >
-              {item.name} - {item.category} - {item.price.toLocaleString()}원
-            </li>
-          ))}
-        </ul>
-      </div>
-    );
-  },
-  (prevProps, nextProps) => {
-    // items가 변경되었을 때만 리렌더링
-    return prevProps.items === nextProps.items;
-  }
-);
-// ComplexForm 컴포넌트
-export const ComplexForm: React.FC = () => {
-  renderLog("ComplexForm rendered");
-  const { addNotification } = useNotification();
-  const [formData, setFormData] = useState({
-    name: "",
-    email: "",
-    age: 0,
-    preferences: [] as string[]
-  });
-
-  const handleSubmit = (e: React.FormEvent) => {
-    e.preventDefault();
-    addNotification("폼이 성공적으로 제출되었습니다", "success");
-  };
-
-  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const { name, value } = e.target;
-    setFormData((prev) => ({
-      ...prev,
-      [name]: name === "age" ? parseInt(value) || 0 : value
-    }));
-  };
-
-  const handlePreferenceChange = (preference: string) => {
-    setFormData((prev) => ({
-      ...prev,
-      preferences: prev.preferences.includes(preference)
-        ? prev.preferences.filter((p) => p !== preference)
-        : [...prev.preferences, preference]
-    }));
-  };
-
-  return (
-    <div className="mt-8">
-      <h2 className="text-2xl font-bold mb-4">복잡한 폼</h2>
-      <form onSubmit={handleSubmit} className="space-y-4">
-        <input
-          type="text"
-          name="name"
-          value={formData.name}
-          onChange={handleInputChange}
-          placeholder="이름"
-          className="w-full p-2 border border-gray-300 rounded text-black"
-        />
-        <input
-          type="email"
-          name="email"
-          value={formData.email}
-          onChange={handleInputChange}
-          placeholder="이메일"
-          className="w-full p-2 border border-gray-300 rounded text-black"
-        />
-        <input
-          type="number"
-          name="age"
-          value={formData.age}
-          onChange={handleInputChange}
-          placeholder="나이"
-          className="w-full p-2 border border-gray-300 rounded text-black"
-        />
-        <div className="space-x-4">
-          {["독서", "운동", "음악", "여행"].map((pref) => (
-            <label key={pref} className="inline-flex items-center">
-              <input
-                type="checkbox"
-                checked={formData.preferences.includes(pref)}
-                onChange={() => handlePreferenceChange(pref)}
-                className="form-checkbox h-5 w-5 text-blue-600"
-              />
-              <span className="ml-2">{pref}</span>
-            </label>
-          ))}
-        </div>
-        <button
-          type="submit"
-          className="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded"
-        >
-          제출
-        </button>
-      </form>
-    </div>
-  );
-};
-
-// NotificationSystem 컴포넌트
-export const NotificationSystem: React.FC = () => {
-  renderLog("NotificationSystem rendered");
-  const { notifications, removeNotification } = useNotification();
-
-  return (
-    <div className="fixed bottom-4 right-4 space-y-2">
-      {notifications.map((notification) => (
-        <div
-          key={notification.id}
-          className={`p-4 rounded shadow-lg ${
-            notification.type === "success"
-              ? "bg-green-500"
-              : notification.type === "error"
-                ? "bg-red-500"
-                : notification.type === "warning"
-                  ? "bg-yellow-500"
-                  : "bg-blue-500"
-          } text-white`}
-        >
-          {notification.message}
-          <button
-            onClick={() => removeNotification(notification.id)}
-            className="ml-4 text-white hover:text-gray-200"
-          >
-            닫기
-          </button>
-        </div>
-      ))}
-    </div>
-  );
-};
-
-// 레이아웃을 위한 새로운 컴포넌트 추가
-const AppLayout = React.memo(({ children }: { children: React.ReactNode }) => {
-  const { theme } = useTheme();
-  return (
-    <div
-      className={`min-h-screen ${theme === "light" ? "bg-gray-100" : "bg-gray-900 text-white"}`}
-    >
-      {children}
-    </div>
-  );
-});
+  ComplexForm,
+  Header,
+  ItemList,
+  NotificationSystem,
+  AppLayout
+} from "./components";
+import { generateItems } from "./utils";
+import { ThemeProvider } from "./contexts/ThemeContext";
+import { AuthProvider } from "./contexts/AuthContext";
+import { NotificationProvider } from "./contexts/NotificationContext";
 
 // 메인 App 컴포넌트
 const App: React.FC = () => {
   const [items, setItems] = useState(generateItems(1000));
 
-  const addItems = useCallback(() => {
+  const addItems = () => {
     setItems((prevItems) => [
       ...prevItems,
       ...generateItems(1000, prevItems.length)
     ]);
-  }, []);
-
-  // ItemList를 메모이제이션
-  const memoizedItemList = useMemo(
-    () => <ItemList items={items} onAddItemsClick={addItems} />,
-    [items, addItems]
-  );
+  };
 
   return (
     <AuthProvider>
       <NotificationProvider>
         <ThemeProvider>
-          {/* <AppContext.Provider value={contextValue}> */}
           <AppLayout>
             <Header />
             <div className="container mx-auto px-4 py-8">
               <div className="flex flex-col md:flex-row">
                 <div className="w-full md:w-1/2 md:pr-4">
-                  {memoizedItemList}
+                  <ItemList items={items} onAddItemsClick={addItems} />
                 </div>
                 <div className="w-full md:w-1/2 md:pl-4">
                   <ComplexForm />
@@ -327,7 +40,6 @@ const App: React.FC = () => {
             </div>
             <NotificationSystem />
           </AppLayout>
-          {/* </AppContext.Provider> */}
         </ThemeProvider>
       </NotificationProvider>
     </AuthProvider>
